+++ mpIIDESY/Mille.h
@@ -55,7 +55,7 @@ class Mille
 
   void mille(int NLC, const float *derLc, int NGL, const float *derGl,
 	     const int *label, float rMeas, float sigma);
-  void special(int nSpecial, const double *floatings, const int *integers);
+  void special(int nSpecial, const float *floatings, const int *integers);
   void kill();
   void end();
 
@@ -69,7 +69,7 @@ class Mille
   /// buffer size for ints and floats
   enum {myBufferSize = 5000};  ///< buffer size for ints and floats
   int   myBufferInt[myBufferSize];   ///< to collect labels etc.
-  double myBufferFloat[myBufferSize]; ///< to collect derivatives etc.
+  float myBufferFloat[myBufferSize]; ///< to collect derivatives etc.
   int   myBufferPos; ///< position in buffer
   bool  myHasSpecial; ///< if true, special(..) already called for this record
   /// largest label allowed: 2^31 - 1
